+++ test/Unitarum/DataBaseTest.php
@@ -44,6 +44,19 @@ class DataBaseTest extends TestCase
         $this->assertEquals($changedEntity, $returnEntity);
     }
 
+    public function testMergeArrayWithoutSecond()
+    {
+        $firstEntity = new User();
+        $firstEntity->setName('Test');
+        $firstEntity->setEmail('test@test.no');
+
+        $secondEntity = null;
+
+        $method = self::getProtectedMethod(DataBase::class, 'mergeArrays');
+        $returnEntity = $method->invokeArgs($this->dataBase, [$firstEntity, $secondEntity]);
+        $this->assertEquals($firstEntity, $returnEntity);
+    }
+
     public function testInsertDataFunctional()
     {
         /* Start transaction */
@@ -76,6 +89,18 @@ class DataBaseTest extends TestCase
         $methodInsert = self::getProtectedMethod(DataBase::class, 'insertData');
         $methodInsert->invokeArgs($this->dataBase, [$insertData, self::TEST_TABLE_USERS]);
 
+        $methodInsert = self::getProtectedMethod(DataBase::class, 'insertData');
+        $methodInsert->invokeArgs($this->dataBase, [$insertData, self::TEST_TABLE_USERS]);
+    }
+
+    /**
+     * @expectedException \Unitarum\Exception\DataBaseException
+     * @expectedExceptionMessageRegExp ~Can not prepare statement! Query:.+ Data:.+~
+     */
+    public function testInsertStatementException()
+    {
+        $this->dataBase->startTransaction();
+        $insertData = [];
         $methodInsert = self::getProtectedMethod(DataBase::class, 'insertData');
         $methodInsert->invokeArgs($this->dataBase, [$insertData, self::TEST_TABLE_USERS]);
     }