+++ CHANGELOG.md
@@ -21,6 +21,11 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ### Security
 - Nothing
 
+## [0.2.1] - 2018-01-20
+- [#6] Check string length in hydrator.
+- [#7] Not marge array without second entity.
+- [#8] Show sql in exception message.
+
 ## [0.2] - 2018-01-19
 
 ### Added